Add initial structure for state managers and controllers
This commit adds a first gameplay scene with some blocks and a
placeholder player object with a rigidbody and two colliders.

Scripts added include simple state managers for the overall game and the
player character in particular. Additionally, the IEntity interface is
added so that the game may track objects that implement this interface
as game entities that have some significance.

.gitignore updated to ignore the Standard Assets directory within
Assets, This allows for importing said standard assets to any
collaborator without needing to add them to the repo needlessly.