Add EnemyBaseController and Change IEntity a bit
Add a function to IEntity
Adds a basic, unmoving enemy to the scene.
Adds EnemyBaseController to the Scripts folder.

EnemyBaseController will be the most basic enemy. Most enemies will
extend from it, and adjust it's methods to fit their own needs. It has
health, attack distances, detection distances and other variables that
can be used by all monsters into the future.

Not sure if this base monster class will also apply to boss monsters,
but probably it will.